sha-crypt: rename `defs` => `consts`
Renames the internal module that defines various constants
