+++ sha-crypt/src/consts.rs
